Consolidated RestClientBase and RestClient classes.
