+++ src/classes/com/urbancode/air/plugin/AppScanSaaS/MAScanRunner.groovy
@@ -1,5 +1,5 @@
 /**
- * © Copyright IBM Corporation 2015.
+ * ï¿½ Copyright IBM Corporation 2015.
  * This is licensed under the following license.
  * The Eclipse Public 1.0 License (http://www.eclipse.org/legal/epl-v10.html)
  * U.S. Government Users Restricted Rights:  Use, duplication or disclosure restricted by GSA ADP Schedule Contract with IBM Corp.
@@ -8,11 +8,11 @@
 package com.urbancode.air.plugin.AppScanSaaS
 
 import java.util.concurrent.TimeUnit;
-import com.urbancode.air.plugin.AppScanSaaS.RestClientBase
+import com.urbancode.air.plugin.AppScanSaaS.RestClient
 import com.urbancode.air.plugin.AppScanSaaS.ScanType
 
 public class MAScanRunner {
-	public static String runMAScan(Properties props, RestClientBase restClient) {
+	public static String runMAScan(Properties props, RestClient restClient) {
 		final def validateReport = false;
 		String issueCountString = "";
 		if (props.containsKey("validateReport")) {
